+++ NAMESPACE
@@ -30,6 +30,7 @@ export(normalizeBySizeFactors)
 export(poolReplicates)
 export(quantifyClustersOlap)
 export(quantifyDivergentLoci)
+export(run_PRIMEloci_example)
 export(subsampleProportion)
 export(subsampleTarget)
 export(summit_decompose)
