+++ man/run_PRIMEloci_example.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/run_PRIMEloci_example.R
+\name{run_PRIMEloci_example}
+\alias{run_PRIMEloci_example}
+\title{Run PRIMEloci on ctss_rse_chr15to17.rds from extdata}
+\usage{
+run_PRIMEloci_example(outdir, python_path = "~/.virtualenvs/prime-env")
+}
+\arguments{
+\item{outdir}{Output directory for the pipeline}
+
+\item{python_path}{Path to Python virtual environment}
+
+\item{num_cores}{Number of CPU cores to use}
+}
+\value{
+GRanges or GRangesList
+}
+\description{
+Run PRIMEloci on ctss_rse_chr15to17.rds from extdata
+}
