--- testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(PRIME)
-
-test_check("PRIME")
